Improve error handling
Streams don't stop after an error such that if you have
pipe streams together you can get multiple errors and
and you'll still get and end event. Refactor so that
we avoid invoking the completion function multiple times.

If you don't specify a completion handler for the
call to get a google file, then if the request
fails you'll get output on stdout.  Add a dummy
function to avoid this.  We currently identify and
report the error in a different way.